feat(metadata): clarify encoding of blob cids
Previously, the code of the Indexing Service always replaces {blob} with a multihash. The previous description implied
occassionally this was a CID, which was incorrect. Moreover, there are cases where providers may
need CID based lookup, so added an additional {blobCID} parameter that can be used to replace with
an actual CID, when one is present in metdata.