feat: add S3-backed datastore implementation (#25)
Moving to go-libstoracha the S3 datastore implementation used by the
storage node (https://github.com/storacha/storage/tree/main/pkg/s3).

I copied it as is. The only change is that I added
`defaults.RemoteCredProvider(*d.Config, d.Handlers)` to the list of
credential providers. The RemoteCredProvider is the one able to resolve
credentials when they are IAM role-based.

Tests are lacking for now, I created an issue to address that when we
have some time: #24.