refactor(testutil): make a public shared testutil package, move in blobindex + bytemap (#40)
# Goals

As we've spun up testing on Guppy, we've once again encountered the
problem of missing generation functions for different common types that
we need for test functions.

This PR accomplishes a major refactor that I've been wanting for a long
time: to move testutil functions in general to a shared package, so we
don't have to keep regenerating them for shared types.

It also finishes an additional endeavor, which is required for Guppy,
that I had hoped to push in a seperate PR, but was just easier to do all
at once: it moves the blob index and byte map packages into
go-libstoracha

# Implementation

- move internal/testutil to testutil
- copy additional functions
- copy the fixtures constants (used in both indexing-service/piri)
- standardize RandomBytes / RandomMultihash to operate the same way
other packages use them
- remove copied internal/testutil sub packages in ipni-publisher and
capabilities
- copy blobindex and bytemap from the indexer

# For discussion

This PR was implemented and tested with corresponding refactors in
indexing-service and piri, which will post seperately

---------

Co-authored-by: Petra Jaros <peeja@peeja.com>