fix: `space/index/add` uses correct reader (#44)
I tried writing a more sensible test that would actually hit this, but I
don't think there's currently a good way to do that. I could force it,
but it's not worth the effort—what we really need is to rework
`go-ucanto`, and that's not a now project. Until then, this makes the
capability work as expected.