fix: add store `Replace` method to allow implementation of conditional writes (#56)
refs https://github.com/ipni/index-provider/issues/454

This PR adds an additional method to `store.Store`:

```go
func Replace(
  ctx context.Context,
  key string,
  oldData io.Reader,
  newLen uint64,
  newData io.Reader,
) error
```

When publishing an advert, IPNI reads the current head, signs an advert,
writes the advert and then writes a new head. If another concurrent
publish writes a new head after the current head is read then the advert
will be lost.

By passing the old bytes to the store the underlying implementation can
generate an `ETag` and send it in an `If-Match` header in order perform
a conditional write that ensures the current bytes of the object equal
the bytes expected.


https://docs.aws.amazon.com/AmazonS3/latest/userguide/conditional-writes.html

In the lambda version of piri we'll implement a datastore that uses the
`If-Match` header and retries writing the advert if it fails the first
time.