feat: add proof data to `pdp/sign/pieces/add` (#67)
This adds links to receipts that prove the node was asked to store the
data it's adding to PDP.

The `blob/accept` receipt will prove that the upload service requested
the blob to be stored (by inspecting the ran invocation details).

The receipt itself includes an effect for `pdp/accept` invocation, the
receipt for which MUST also be included. The `pdp/accept` receipt
provides the inclusion proof that a piece was included in an aggregate.

Assuming we trust the upload service to be sending `blob/accept`
invocations, the only thing the signing service needs to do is verify
that the piece CIDs correspond to the blob CIDs from the invocations.

refs https://github.com/storacha/piri-signing-service/issues/2